requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.32
tomli==2.0.1
urllib3==1.26.9
vk-api==11.9.7
